requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asn1crypto==1.5.1
asyncio==3.4.3
attrs==23.1.0
auto-py-to-exe==2.43.3
beautifulsoup4==4.12.2
bech32==1.2.0
bip-utils==2.9.3
bitarray==2.9.2
bottle==0.12.25
bottle-websocket==0.2.9
captchatools==1.5.0
cbor2==5.6.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.0
coincurve==19.0.1
colorama==0.4.6
cosmospy==6.0.0
cosmospy_protobuf==0.4.0
crcmod==1.7
cx_Freeze==7.0.0
cx_Logging==3.2.0
cytoolz==0.12.3
dateparser==1.2.0
ecdsa==0.19.0
ed25519-blake2b==1.4.1
Eel==0.16.0
et-xmlfile==1.1.0
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.1.0
eth-utils==4.1.0
eth_abi==5.1.0
frozenlist==1.4.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
hdwallets==0.1.2
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lief==0.14.1
lru-dict==1.2.0
mnemonic==0.21
mospy-wallet @ git+https://github.com/ctrl-Felix/mospy.git@ea07705b6ec4c76cc355b7dc933fae7c09fd8429
multidict==6.0.4
Nuitka==2.2.1
numpy==1.26.4
openpyxl==3.1.2
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parsimonious==0.10.0
pefile==2023.2.7
Pillow==10.1.0
playwright==1.43.0
protobuf==5.26.1
py-sr25519-bindings==0.2.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pyee==11.1.0
pygame==2.5.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
PyNaCl==1.5.0
pyparsing==3.1.2
pyperclip==1.8.2
pyppeteer==2.0.0
PySocks==1.7.1
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-capsolver==0.9.2
pytz==2023.3.post1
pyunormalize==15.1.0
pywin32==306
pywin32-ctypes==0.2.2
referencing==0.35.0
regex==2023.10.3
requests==2.31.0
rlp==4.0.0
rpds-py==0.18.0
safe-pysha3==1.0.4
selenium==4.16.0
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tenacity==8.2.3
toolz==0.12.1
tqdm==4.66.2
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
urllib3==1.26.18
web3==6.17.2
webdriver-manager==4.0.1
websockets==10.4
wheel==0.43.0
whichcraft==0.6.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
zope.event==5.0
zope.interface==6.3
zstandard==0.22.0
